last step before removing logic from frontend
