+++ server/DispatchGame.js
@@ -1,4 +1,3 @@
-
 const {PHASE, REASON, STATUS, DECK} = require('./constants.js');
 
 
@@ -7,14 +6,14 @@ const getClientState = (game, {playerIndex}) => {
         cards: [game.cards.map( (card) => {
             return {
                 word: card.word,
-                status: card.revealed[playerNumner]
+                status: card.revealed[playerIndex]
             }
         })],
         history: [...game.history],
         current_turn: {...game.current_turn},
         nessage: game.message,
         game_over: game.game_over,
-        validationErrors: game.validationErrors[playerIndex],
+        validationError: game.validationErrors[playerIndex],
         canClick: game.canClick({playerIndex}),
         canClue: game.canGiveClue({playerIndex}),
         canPass: game.canPass({playerIndex})
@@ -23,21 +22,32 @@ const getClientState = (game, {playerIndex}) => {
     return clientState;
 }
 
+const getClientStates = (game, success) => {
+  return {
+    success: true,
+    clientStates: [getClientState(game, 0), getClientState(game, 1)]
+  }
+}
+
 
 const dispatchClickCard = (game, {playerIndex, cardIndex}) => {
-    return 0;
+    const {success} = game.clickCardNumber({playerIndex, cardIndex});
+    return getClientStates(game, success);
 }
 
 const dispatchClickPass = (game, {playerIndex}) => {
-    return 0;
+    const {success} = game.clickPass({playerIndex});
+    return getClientStates(game, success);
 }
 
 const dispatchSendClue = (game, {playerIndex, clue, number}) => {
-    const {success, message} = game.sendClue(playerIndex, clue, number);
-
-    return 0;
+    const {success} = game.sendClue(playerIndex, clue, number);
+    return getClientStates(game, success);
 }
 
 const dispatchRefresh = (game, {playerIndex}) => {
-    return 
+    return getClientStates(game, true);
 }
+
+
+module.exports = {dispatchClickCard, dispatchClickPass, dispatchSendClue, dispatchRefresh};
